feat: add product models and configure handlers

+++ src/models/product_models.rs
@@ -1,16 +1,17 @@
 use serde::{Deserialize, Serialize};
+use rust_decimal::Decimal;
 
 #[derive(Debug, Deserialize)]
 pub struct CrearProducto {
     pub nombre: String,
-    pub precio: f64,
+    pub precio: Decimal,
     pub stock: i32,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct ActualizarProducto {
     pub nombre: String,
-    pub precio: f64,
+    pub precio: Decimal,
     pub stock: i32,
 }
 
@@ -18,7 +19,7 @@ pub struct ActualizarProducto {
 pub struct Producto {
     pub id: i32,
     pub nombre: String,
-    pub precio: f64,
+    pub precio: Decimal,
     pub stock: i32,
     pub creado_el: String,
 }
